certtool: lifted limits on file size to load

+++ src/certtool-common.h
@@ -111,8 +111,8 @@ const char *get_password(common_info_st * cinfo, unsigned int *flags,
 			 int confirm);
 
 extern unsigned char *lbuffer;
-extern int lbuffer_size;
+extern unsigned long lbuffer_size;
 
-void fix_lbuffer(unsigned);
+void fix_lbuffer(unsigned long);
 
 #endif
